refactor: waiting-room 외부 참조 경로 정리
